requirements.txt
PyPDF2
streamlit
streamlit-option-menu
pandas
python-docx
